Rename folder - Work in progress: implementing validation for IBM1068I
